refactor: Delete old file
