--- src/utils/mobileClient.js
@@ -1,35 +1,0 @@
-import { getDeviceName } from 'cozy-device-helper'
-
-export const getScope = m => {
-  if (m.permissions === undefined) {
-    throw new Error(`Your manifest must have a 'permissions' key.`)
-  }
-
-  return Object.keys(m.permissions).map(permission => {
-    const { type /*, verbs, selector, values*/ } = m.permissions[permission]
-
-    return type
-  })
-}
-
-export const getSoftwareName = m => {
-  if (m.name === undefined) {
-    throw new Error(`Your manifest must have a 'name' key.`)
-  }
-
-  return m.name_prefix ? `${m.name_prefix} ${m.name}` : m.name
-}
-
-export const getClientName = m => `${getSoftwareName(m)} (${getDeviceName()})`
-
-export const getManifestOptions = manifest => {
-  const cozyPolicyURI = 'https://files.cozycloud.cc/cgu.pdf'
-
-  return {
-    oauth: {
-      clientName: getClientName(manifest),
-      policyURI: manifest.name_prefix === 'Cozy' ? cozyPolicyURI : ''
-    },
-    scope: getScope(manifest)
-  }
-}
